Bug fix - operations after '=' on leftover values
Allows users for normal flow, like a normal calculator.
Let's say you've pressed 3 + 6 and pressed =, the result
is 9, its stored in memory accumulator. Next time you
press 1 + 2, it previously operated on previous result
(accumulator value) and the newest displayValue (2).
We fix this problem by updating the accumulator value to
new value (1) as soon as we see that previous operator
was '='