Added the feature start new operation after '='
COULDN'T do 88 + 20 = 108 + 2 = 110.
it shows 88 + 20 = 108 and then 20 + (108) + 2 = 110
- that is, now, you can directly start operating after
doing one operation with '='.
Take a + b = X and c + d = Z
This was possible because before, the operated value
(before) (X) directly went to accumulator and when new
operation was done (c + d), the accumulator was not empty
and hence, the operation was done with the previous value
in accumulator (X). Now, we update accumulator if we get
a new operator (+) to have the new value (c).
Other operations proceed as normal.