Solve '.' press problem after memory operation
When operation was done and . was pressed, it was
stuck there until the next operation and in next number
press, it showed the number only.
eg: 8 + 3 = (shows 11) --> 11 in memory.accumulator
Now, if you pressed '.' before, it showed 11. and again
if you pressed 3, it showed only 3.
i.e. 11 . --> 11. 3 --> 3
This happens because operation pachi displayBuffer hasn't
been refreshed to false.
Solved by chaning the logic from inside
if(numKeys.includes(press.target)) to outside it,
so that it happens for every keypress an not just for
numKeys in the next operation after operation.