Solve the continuous operation problem by commit-1
Previous commit took the condition on 101 line
outside which meant that displayValue was cleared
after any keyPress after operation.
We wanted that to solve the '.' problem. But doing
that meant that when we pressed (say) + after =,
the displayValue would be cleared. But the later
operations depend on displayValue.
So, the solution was to check if the value is '.' and
if the previous button was a operation(using displayBuffer)
[SEPERATELY] for the previous commit problem