Make the digits size decrease with display
We do this by making the digits and display both
take the units of viewport width(vw), because it
means that when the width of device decreases,
both of the display and the font inside the display
decrease in size.
We can use any relative sizing here, but we MUST
use the relative sizing that changes with display
(eg: vw).